Use subcommands and add --help and --version
